requirements.txt
a2a-sdk[http-server]>=0.3.25
httpx>=0.28.1
uvicorn>=0.42.0
